# openapi/go-api.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "\nTraceRTM is a comprehensive requirements traceability management system that helps teams\ntrack relationships between requirements, features, and tasks throughout the development lifecycle.\n\n## Features\n- Project and Item Management\n- Link Management for traceability\n- AI Agent Coordination\n- Graph-based relationship analysis\n- Full-text search with pgvector\n- Real-time updates via WebSocket and NATS\n\n## Authentication\nCurrently supports WorkOS AuthKit. Set Authorization header with Bearer token.",
    "title": "TraceRTM API",
    "contact": {
      "name": "TraceRTM Support",
      "url": "https://github.com/kooshapari/tracertm-backend",
      "email": "support@tracertm.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/api/v1/projects/{projectId}/versions/compare": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns the differences between two project versions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "versions"
        ],
        "summary": "Compare two versions",
        "parameters": [
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Source version ID",
            "name": "from",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Target version ID",
            "name": "to",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Diff result",
            "schema": {
              "$ref": "#/definitions/handlers.CompareVersionsResponse"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "404": {
            "description": "Version not found",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/versions/compare/bulk": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns differences for multiple version pairs in a single request",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "versions"
        ],
        "summary": "Compare multiple version pairs",
        "parameters": [
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "description": "Comparison pairs",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handlers.BulkCompareVersionsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bulk diff results",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/versions/compare/summary": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns a high-level summary of changes between two versions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "versions"
        ],
        "summary": "Get summary of version differences",
        "parameters": [
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Source version ID",
            "name": "from",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Target version ID",
            "name": "to",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Diff summary",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "404": {
            "description": "Version not found",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "handlers.BulkCompareVersionsRequest": {
      "type": "object",
      "required": [
        "comparisons"
      ],
      "properties": {
        "comparisons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/handlers.VersionComparisonPair"
          }
        }
      }
    },
...
    "handlers.CompareVersionsResponse": {
      "type": "object",
      "properties": {
        "diff": {
          "$ref": "#/definitions/temporal.VersionDiff"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "handlers.ErrorResponse": {
      "description": "Error response with error message",
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
...
    "handlers.VersionComparisonPair": {
      "type": "object",
      "required": [
        "fromVersionId",
        "toVersionId"
      ],
      "properties": {
        "fromVersionId": {
          "type": "string"
        },
        "toVersionId": {
          "type": "string"
        }
      }
    },
...
    "temporal.DiffChangeType": {
      "type": "string",
      "enum": [
        "added",
        "removed",
        "modified"
      ],
      "x-enum-varnames": [
        "DiffChangeTypeAdded",
        "DiffChangeTypeRemoved",
        "DiffChangeTypeModified"
      ]
    },
    "temporal.DiffItem": {
      "type": "object",
      "properties": {
        "changeType": {
          "$ref": "#/definitions/temporal.DiffChangeType"
        },
        "fieldChanges": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/temporal.FieldChange"
          }
        },
        "itemId": {
          "type": "string"
        },
        "significance": {
          "$ref": "#/definitions/temporal.DiffSignificance"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "temporal.DiffSignificance": {
      "type": "string",
      "enum": [
        "minor",
        "moderate",
        "major",
        "breaking"
      ],
      "x-enum-varnames": [
        "DiffSignificanceMinor",
        "DiffSignificanceModerate",
        "DiffSignificanceMajor",
        "DiffSignificanceBreaking"
      ]
    },
    "temporal.DiffStatistics": {
      "type": "object",
      "properties": {
        "addedCount": {
          "type": "integer"
        },
        "modifiedCount": {
          "type": "integer"
        },
        "removedCount": {
          "type": "integer"
        },
        "totalChanges": {
          "type": "integer"
        },
        "unchangedCount": {
          "type": "integer"
        }
      }
    },
    "temporal.FieldChange": {
      "type": "object",
      "properties": {
        "changeType": {
          "description": "added, removed, modified",
          "type": "string"
        },
        "field": {
          "type": "string"
        },
        "newValue": {},
        "oldValue": {}
      }
    },
    "temporal.VersionDiff": {
      "type": "object",
      "properties": {
        "added": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/temporal.DiffItem"
          }
        },
        "computedAt": {
          "type": "string"
        },
        "modified": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/temporal.DiffItem"
          }
        },
        "removed": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/temporal.DiffItem"
          }
        },
        "stats": {
          "$ref": "#/definitions/temporal.DiffStatistics"
        },
        "unchanged": {
          "type": "integer"
        },
        "versionA": {
          "type": "string"
        },
        "versionANumber": {
          "type": "integer"
        },
        "versionB": {
          "type": "string"
        },
        "versionBNumber": {
          "type": "integer"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "Enter your bearer token in the format: Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}